Split gbuffer
