+++ common/gbuffer.cpp
@@ -44,255 +44,12 @@
  * - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 
 #include "gbuffer.h"
-#include "video.h"
-#include "drawbuff.h"
-#include "ww_win.h"
-#include "font.h"
-
-#include <cstdlib>
-#include <algorithm>
 
 int TotalLocks = 0;
+extern bool GameInFocus;
 
-/*=========================================================================*/
-/* The following PRIVATE functions are in this file:                       */
-/*=========================================================================*/
-
-/*= = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = = =*/
-
-/***************************************************************************
- * GVPC::GRAPHICVIEWPORTCLASS -- Constructor for basic view port class     *
- *                                                                         *
- * INPUT:      GraphicBufferClass * gbuffer - buffer to attach to          *
- *             int x                        - x offset into buffer         *
- *             int y                        - y offset into buffer         *
- *             int w                        - view port width in pixels    *
- *             int h                        - view port height in pixels   *
- *                                                                         *
- * OUTPUT:     Constructors may not have a return value                    *
- *                                                                         *
- * HISTORY:                                                                *
- *   05/09/1994 PWG : Created.                                             *
- *=========================================================================*/
-GraphicViewPortClass::GraphicViewPortClass(GraphicBufferClass* gbuffer, int x, int y, int w, int h)
-    : Offset(0)
-    , Width(0)
-    , Height(0)
-    , XAdd(0)
-    , XPos(0)
-    , YPos(0)
-    , Pitch(0)
-    , GraphicBuff(nullptr)
-    , IsHardware(false)
-    , LockCount(0)
-{
-    Attach(gbuffer, x, y, w, h);
-}
-
-/***************************************************************************
- * GVPC::GRAPHICVIEWPORTCLASS -- Default constructor for view port class   *
- *                                                                         *
- * INPUT:      none                                                        *
- *                                                                         *
- * OUTPUT:     none                                                        *
- *                                                                         *
- * HISTORY:                                                                *
- *   05/09/1994 PWG : Created.                                             *
- *=========================================================================*/
-GraphicViewPortClass::GraphicViewPortClass()
-    : Offset(0)
-    , Width(0)
-    , Height(0)
-    , XAdd(0)
-    , XPos(0)
-    , YPos(0)
-    , Pitch(0)
-    , GraphicBuff(nullptr)
-    , IsHardware(false)
-    , LockCount(0)
-{
-}
-
-/***************************************************************************
- * GVPC::~GRAPHICVIEWPORTCLASS -- Destructor for GraphicViewPortClass      *
- *                                                                         *
- * INPUT:      none                                                        *
- *                                                                         *
- * OUTPUT:     A destructor may not return a value.                        *
- *                                                                         *
- * HISTORY:                                                                *
- *   05/10/1994 PWG : Created.                                             *
- *=========================================================================*/
-GraphicViewPortClass::~GraphicViewPortClass(void)
-{
-}
-
-/***************************************************************************
- * GVPC::ATTACH -- Attaches a viewport to a buffer class                   *
- *                                                                         *
- * INPUT:    GraphicBufferClass *g_buff  -  pointer to gbuff to attach to  *
- *           int x                       - x position to attach to         *
- *           int y                       - y position to attach to         *
- *           int w                       - width of the view port          *
- *           int h                       - height of the view port         *
- *                                                                         *
- * OUTPUT:     none                                                        *
- *                                                                         *
- * HISTORY:                                                                *
- *   05/10/1994 PWG : Created.                                             *
- *=========================================================================*/
-void GraphicViewPortClass::Attach(GraphicBufferClass* gbuffer, int x, int y, int w, int h)
-{
-    /*======================================================================*/
-    /* Can not attach a Graphic View Port if it is actually the physical    */
-    /*     representation of a Graphic Buffer.                              */
-    /*======================================================================*/
-    if (this == Get_Graphic_Buffer()) {
-        return;
-    }
-
-    /*======================================================================*/
-    /* Verify that the x and y coordinates are valid and placed within the  */
-    /*      physical buffer.                                                */
-    /*======================================================================*/
-    if (x < 0)                         // you cannot place view port off
-        x = 0;                         //   the left edge of physical buf
-    if (x >= gbuffer->Get_Width())     // you cannot place left edge off
-        x = gbuffer->Get_Width() - 1;  //   the right edge of physical buf
-    if (y < 0)                         // you cannot place view port off
-        y = 0;                         //   the top edge of physical buf
-    if (y >= gbuffer->Get_Height())    // you cannot place view port off
-        y = gbuffer->Get_Height() - 1; //   bottom edge of physical buf
-
-    /*======================================================================*/
-    /* Adjust the width and height of necessary                             */
-    /*======================================================================*/
-    if (x + w > gbuffer->Get_Width()) // if the x plus width is larger
-        w = gbuffer->Get_Width() - x; //    than physical, fix width
-
-    if (y + h > gbuffer->Get_Height()) // if the y plus height is larger
-        h = gbuffer->Get_Height() - y; //   than physical, fix height
-
-    /*======================================================================*/
-    /* Get a pointer to the top left edge of the buffer.                    */
-    /*======================================================================*/
-    Offset = gbuffer->Get_Offset() + ((gbuffer->Get_Width() + gbuffer->Get_Pitch()) * y) + x;
-
-    /*======================================================================*/
-    /* Copy over all of the variables that we need to store.                */
-    /*======================================================================*/
-    XPos = x;
-    YPos = y;
-    XAdd = gbuffer->Get_Width() - w;
-    Width = w;
-    Height = h;
-    Pitch = gbuffer->Get_Pitch();
-    GraphicBuff = gbuffer;
-    IsHardware = gbuffer->IsHardware;
-}
-
-/***************************************************************************
- * GVPC::CHANGE -- Changes position and size of a Graphic View Port        *
- *                                                                         *
- * INPUT:       int the new x pixel position of the graphic view port      *
- *              int the new y pixel position of the graphic view port      *
- *              int the new width of the viewport in pixels                *
- *              int the new height of the viewport in pixels               *
- *                                                                         *
- * OUTPUT:      BOOL whether the Graphic View Port could be sucessfully    *
- *                      resized.                                           *
- *                                                                         *
- * WARNINGS:   You may not resize a Graphic View Port which is derived     *
- *                      from a Graphic View Port Buffer,                   *
- *                                                                         *
- * HISTORY:                                                                *
- *   09/14/1994 SKB : Created.                                             *
- *=========================================================================*/
-bool GraphicViewPortClass::Change(int x, int y, int w, int h)
-{
-    /*======================================================================*/
-    /* Can not change a Graphic View Port if it is actually the physical    */
-    /*      representation of a Graphic Buffer.                             */
-    /*======================================================================*/
-    if (this == Get_Graphic_Buffer()) {
-        return false;
-    }
-
-    /*======================================================================*/
-    /* Since there is no allocated information, just re-attach it to the    */
-    /*      existing graphic buffer as if we were creating the              */
-    /*      GraphicViewPort.                                                */
-    /*======================================================================*/
-    Attach(Get_Graphic_Buffer(), x, y, w, h);
-    return true;
-}
-
-/***************************************************************************
- * Draw_Rect -- Draws a rectangle to the LogicPage.                        *
- *                                                                         *
- *    This routine will draw a rectangle to the LogicPage.  The rectangle  *
- *    doesn't have to be aligned on the vertical or horizontal axis.  In   *
- *    fact, it doesn't even have to be a rectangle.  The "square" can be   *
- *    skewed.                                                              *
- *                                                                         *
- * INPUT:   x1_pixel, y1_pixel   -- One corner.                            *
- *                                                                         *
- *          x2_pixel, y2_pixel   -- The other corner.                      *
- *                                                                         *
- *          color                -- The color to draw the lines.           *
- *                                                                         *
- * OUTPUT:  none                                                           *
- *                                                                         *
- * WARNINGS:   None, but the rectangle will be clipped to the current      *
- *             draw line clipping rectangle.                               *
- *                                                                         *
- * HISTORY:                                                                *
- *   08/20/1993 JLB : Created.                                             *
- *=========================================================================*/
-void GraphicViewPortClass::Draw_Rect(int x1_pixel, int y1_pixel, int x2_pixel, int y2_pixel, unsigned char color)
-{
-    Lock();
-    Draw_Line(x1_pixel, y1_pixel, x2_pixel, y1_pixel, color);
-    Draw_Line(x1_pixel, y2_pixel, x2_pixel, y2_pixel, color);
-    Draw_Line(x1_pixel, y1_pixel, x1_pixel, y2_pixel, color);
-    Draw_Line(x2_pixel, y1_pixel, x2_pixel, y2_pixel, color);
-    Unlock();
-}
-
-/***************************************************************************
- * GVPC::DRAW_STAMP -- stub function to draw a tile on a graphic view port *
- *                     This version clips the tile to a window             *
- *                                                                         *
- * INPUT:                                                                  *
- *                                                                         *
- * OUTPUT:                                                                 *
- *                                                                         *
- * WARNINGS:                                                               *
- *                                                                         *
- * HISTORY:                                                                *
- *    07/31/1995 BWG : Created.                                            *
- *=========================================================================*/
-void GraphicViewPortClass::Draw_Stamp(void const* icondata,
-                                      int icon,
-                                      int x_pixel,
-                                      int y_pixel,
-                                      void const* remap,
-                                      int clip_window)
-{
-    if (Lock()) {
-        Buffer_Draw_Stamp_Clip(this,
-                               icondata,
-                               icon,
-                               x_pixel,
-                               y_pixel,
-                               remap,
-                               WindowList[clip_window][WINDOWX],
-                               WindowList[clip_window][WINDOWY],
-                               WindowList[clip_window][WINDOWWIDTH],
-                               WindowList[clip_window][WINDOWHEIGHT]);
-    }
-    Unlock();
-}
+extern void Block_Mouse(GraphicBufferClass* buffer);
+extern void Unblock_Mouse(GraphicBufferClass* buffer);
 
 /***************************************************************************
  * GBC::DD_INIT -- Inits a direct draw surface for a GBC                   *
@@ -493,61 +250,6 @@ GraphicBufferClass::~GraphicBufferClass()
     Un_Init();
 }
 
-/***************************************************************************
- * SET_LOGIC_PAGE -- Sets LogicPage to new buffer                          *
- *                                                                         *
- * INPUT:      GraphicBufferClass * the buffer we are going to set         *
- *                                                                         *
- * OUTPUT:     GraphicBufferClass * the previous buffer type               *
- *                                                                         *
- * WARNINGS:                                                               *
- *                                                                         *
- * HISTORY:                                                                *
- *   02/23/1995 PWG : Created.                                             *
- *=========================================================================*/
-GraphicViewPortClass* Set_Logic_Page(GraphicViewPortClass* ptr)
-{
-    GraphicViewPortClass* old = LogicPage;
-    LogicPage = ptr;
-    return (old);
-}
-
-/***************************************************************************
- * SET_LOGIC_PAGE -- Sets LogicPage to new buffer                          *
- *                                                                         *
- * INPUT:      GraphicBufferClass & the buffer we are going to set         *
- *                                                                         *
- * OUTPUT:     GraphicBufferClass * the previous buffer type               *
- *                                                                         *
- * WARNINGS:                                                               *
- *                                                                         *
- * HISTORY:                                                                *
- *   02/23/1995 PWG : Created.                                             *
- *=========================================================================*/
-GraphicViewPortClass* Set_Logic_Page(GraphicViewPortClass& ptr)
-{
-    GraphicViewPortClass* old = LogicPage;
-    LogicPage = &ptr;
-    return (old);
-}
-
-/***************************************************************************
- * GBC::LOCK -- Locks a Direct Draw Surface                                *
- *                                                                         *
- * INPUT:      none                                                        *
- *                                                                         *
- * OUTPUT:     none                                                        *
- *                                                                         *
- * HISTORY:                                                                *
- *   10/09/1995     : Created.                                             *
- *   10/09/1995     : Code stolen from Steve Tall                          *
- *=========================================================================*/
-extern void Colour_Debug(int call_number);
-extern bool GameInFocus;
-
-extern void Block_Mouse(GraphicBufferClass* buffer);
-extern void Unblock_Mouse(GraphicBufferClass* buffer);
-
 bool GraphicBufferClass::Lock()
 {
     //
@@ -656,52 +358,4 @@ bool GraphicBufferClass::IsAllocated() const
     }
 
     return VideoSurfacePtr->IsAllocated();
-}
-
-/***********************************************************************************************
- * GVPC::DD_Linear_Blit_To_Linear -- blit using the hardware blitter                           *
- *                                                                                             *
- *                                                                                             *
- *                                                                                             *
- * INPUT:    destination vvpc                                                                  *
- *           x coord to blit from                                                              *
- *           y coord to blit from                                                              *
- *           x coord to blit to                                                                *
- *           y coord to blit to                                                                *
- *           width to blit                                                                     *
- *           height to blit                                                                    *
- *                                                                                             *
- * OUTPUT:   DD_OK if successful                                                               *
- *                                                                                             *
- * WARNINGS: None                                                                              *
- *                                                                                             *
- * HISTORY:                                                                                    *
- *   09-22-95 11:05am ST : Created                                                             *
- *=============================================================================================*/
-
-bool GraphicViewPortClass::DD_Linear_Blit_To_Linear(GraphicViewPortClass& dest,
-                                                    int source_x,
-                                                    int source_y,
-                                                    int dest_x,
-                                                    int dest_y,
-                                                    int width,
-                                                    int height,
-                                                    bool mask)
-
-{
-    Rect source_rectangle = {0};
-    source_rectangle.X = source_x;
-    source_rectangle.Y = source_y;
-    source_rectangle.Width = width;
-    source_rectangle.Height = height;
-
-    Rect dest_rectangle = {0};
-    dest_rectangle.X = dest_x;
-    dest_rectangle.Y = dest_y;
-    dest_rectangle.Width = width;
-    dest_rectangle.Height = height;
-
-    dest.GraphicBuff->Get_DD_Surface()->Blt(dest_rectangle, GraphicBuff->Get_DD_Surface(), source_rectangle, mask);
-
-    return true;
 }
